Move all error codes in a module `errors.rs`
